+++ Ya/Algebra/Abstract.hs
@@ -30,8 +30,8 @@ newtype U_T_I_II t u i ii = U_T_I_II (u (t i) ii)
 
 newtype U_I_UU_II_III u uu i ii iii = U_I_UU_II_III (u i (uu ii iii))
 
-newtype U_UU_UUU_V_I_II_III_UUUU u uu uuu uuuu v i ii iii = U_UU_UUU_V_I_II_III_UUUU
-	(u (uu (v uuu i ii) (v uuu i iii)) (v uuuu i (uuu ii iii)))
+newtype U_UU_UUU_V_III_I_II_UUUU u uu uuu uuuu v i ii x = U_UU_UUU_V_III_I_II_UUUU
+	(u (uu (v uuu x i) (v uuu x ii)) (v uuu x (uuuu i ii)))
 
 newtype U_V_UU_UUU_UUUU_T_TT_I_II_III u v uu uuu uuuu t tt i ii iii =
 	U_V_UU_UUU_UUUU_T_TT_I_II_III (u (uuu (t i) (tt ii)) (v uu (uuuu i ii) iii))
@@ -58,6 +58,10 @@ type Flat = U_I_II
 
 type Dual = U_II_I
 
+type family Flip v where
+	Flip Flat = Dual
+	Flip Dual = Flat
+
 type family Supertype e where
 	Supertype (I i) = i
 	Supertype (U_1_II u i ii) = u Unit ii
@@ -72,8 +76,8 @@ type family Supertype e where
 	Supertype (U_I_T_II t u i ii) = u i (t ii)
 	Supertype (U_T_I_II t u i ii) = u (t i) ii
 	Supertype (U_I_UU_II_III u uu i ii iii) = u i (uu ii iii)
-	Supertype (U_UU_UUU_V_I_II_III_UUUU u uu uuu uuuu v i ii iii) = 
-		u (uu (v uuu i ii) (v uuu i iii)) (v uuuu i (uuu ii iii))
+	Supertype (U_UU_UUU_V_III_I_II_UUUU u uu uuu uuuu v i ii iii) = 
+		u (uu (v uuu iii i) (v uuu iii ii)) (v uuu iii (uuuu i ii))
 	Supertype (U_V_UU_UUU_UUUU_T_TT_I_II_III u v uu uuu uuuu t tt i ii iii) =
 		u (uuu (t i) (tt ii)) (v uu (uuuu i ii) iii)
 	Supertype (UU_V_U_I_II_T_II v u uu t i ii) = uu (v u ii i) (t i)
@@ -154,11 +158,11 @@ instance Castable Dual Arrow (U_I_UU_II_III u uu i ii iii)
 instance Castable Flat Arrow (U_I_UU_II_III u uu i ii iii)
 	where cast = U_I_II (\(U_I_UU_II_III x) -> x)
 
-instance Castable Flat Arrow (U_UU_UUU_V_I_II_III_UUUU u uu uuu uuuu v i ii iii)
-	where cast = U_I_II (\(U_UU_UUU_V_I_II_III_UUUU x) -> x)
+instance Castable Flat Arrow (U_UU_UUU_V_III_I_II_UUUU u uu uuu uuuu v i ii iii)
+	where cast = U_I_II (\(U_UU_UUU_V_III_I_II_UUUU x) -> x)
 
-instance Castable Dual Arrow (U_UU_UUU_V_I_II_III_UUUU u uu uuu uuuu v i ii iii)
-	where cast = U_II_I U_UU_UUU_V_I_II_III_UUUU 
+instance Castable Dual Arrow (U_UU_UUU_V_III_I_II_UUUU u uu uuu uuuu v i ii iii)
+	where cast = U_II_I U_UU_UUU_V_III_I_II_UUUU
 
 instance Castable Flat Arrow (U_V_UU_UUU_UUUU_T_TT_I_II_III u v uu uuu uuuu t tt i ii iii_)
 	where cast = U_I_II (\(U_V_UU_UUU_UUUU_T_TT_I_II_III x) -> x)