+++ CHANGELOG.md
@@ -163,3 +163,5 @@
 * Define `point` expression
 * Define `TT_T_I` data type
 * Define `U_UU_UUU_V_I_II_III_UUUU` data type
+* Define `Flip` type family
+* Define `Order` type family